部门ID统一为long

+++ src/main/java/cn/felord/domain/contactbook/tag/TagUserActionResponse.java
@@ -14,6 +14,6 @@ import java.util.List;
 @Data
 public class TagUserActionResponse extends WeComResponse {
      private String invalidlist;
-     private List<Integer> invalidparty;
+     private List<Long> invalidparty;
 
 }
